Extract business logic from ContractDropdown, make it completely decoupled
